removes mention of vagrant-cachier from develop.md
The vagrant-cachier plugin requires that all reboots to
Vagrant-managed VMs be done via `vagrant reload`. Since
the Ansible playbooks reboot as part of the provisioning
run, cachier doesn't work right. After 8 months of broken
functionality, it's time to cull this plugin.

See here for discussion of the reboot problem:
https://github.com/fgrehm/vagrant-cachier/issues/128

At a later date it may be wise to virtualize the apt repo
and mirror packages for easy installation.