moves host_vars to to group_vars
Using group_vars vastly simplifies the environment tiers
(i.e. development, staging, and prod). It will, however,
require changes to the prod-only "inventory" file, which
should be handled with care. It's also possible to handle
the group membership for the prod hosts in a task as part
of the prod-specific playbook, which may be useful as a failsafe.