Removes unnecessary config lines from Rakefile
No testing environment uses a sudo password, so we can ditch that block.
The `vagrant up` command adds several seconds to each spec test run, and
runs every time regardless. That's foolish and slow. Assume that if the
spectests are being run, the hosts are up. The snap_ci scripts already
take care of creating the hosts before testing.