Downloads OSSEC base packages in Snap-CI staging environment
Unfortunately we can't upload the ossec-agent and ossec-server deb files
to the Snap-CI platform as static assets used in the build. The packages
are built in an external repo, so we can't trivially automate building
them, either. Since the files only change once every few months, it's OK
to pull them down based on what's in the FPF repo and use that.

For local build in VirtualBox, the testing scripts still expect the
packages to exist locally. If not found, the install-local-packages role
will throw an error and display an informative error message pointing to
the docs.

In both cases (Snap-CI and local development), the securedrop-*
metapackages are rebuilt and reinstalled on every staging provisioning
run.