KITCHEN: Adds .kitchen.yml config file for test-kitchen
Uses multiple suites, but inappropriately assigning
app and mon hosts to the "build" suite.

Adds custom vars for ossec deb file locations in CI

The role default vars for importing the DEBIAN control files and
post-install hooks is relative to the role path. That's good enough when
building in the build VM via Vagrant, but breaks portability for the
role, which is necessary for testing. Adding vars overrides solves the
problem.

Adds securedrop extra files dir to .kitchen.yml