Uses synchronize module to copy app-code package files
Adjusts relative paths in app-code build role with careful attention to
trailing slashes, since the synchronize module is just a wrapper around
rsync. Will need to be attentive to permissions and ownership as well.

Also creates default var for build directory path.
The app-code building role was previously using a hard-coded relative
path (`../../securedrop`) for finding the app code. That makes the role
less portable than it could be, which matters because CI testing may
isolate the role. So we'll still default to that relative path, but
allow overriding via a var.