Creates separate plays in build playbook for debs
Separating the role the includes for each deb package into discrete
plays. This is necessary because we're using wrapper roles to manage the
builds, and Ansible dependency resolution for roles conflates child and
parent roles. The problem has been patched in Ansible core and will be
released as part of 2.0.

Note that specifying `allow_duplicates: yes` in the children roles does
not solve the problem. We're already setting that attribute in the
parent role.