Adds support for static files to build-debian-package role
Already synchronizing directories of static files, but doesn't hurt to
add support for custom filepaths to `copy` in. Requires "src" and "dest"
attributes for the filepaths provided in a list. This is mostly
necessary due to the technical debt of storing the DEBIAN control files
for the built packages in a higher-level directory in the securedrop git
repository.

Also carefully manipulates the rsync command used by `synchronize` to
ensure correct permissions and ownership on the dest files.