Updates Bash script for Prod CI stage
Changes:

  * Filters out staging vars such as `install_local_packages`
  * Checks environment variables (e.g. "CI")
  * Ensures prod hosts (not staging) are destroyed in EXIT trap
  * Reduce sleep time after rebooting prod hosts

The sleep reduction from 300 to 180 seconds isn't tested, but should be
enough time for the ATHS for SSH to come up on the remote machines.
That's all we need to wait for. Don't forget the CI playbook runs prior
to the sleep, so we're still waiting about 300 seconds altogether for
the prod hosts to come up.