Serverspec: Adds deb package contents checks for OSSEC packages
Stringent regression tests for the file contents of the Debian packages
created by the build VM. Inspects package contents post-build and
ensures:

  * matching ownership (root/root)
  * matching permissions
  * exact file list

If any files are added or removed to the Debian packages, these tests
will need to be updated. This fail-fast approach is intended, and will
catch a variety of surprising regressions, particularly as the build
scripts are refactored.

The `dpkg -c` output was captured from the 0.3.5 versions of the deb
packages, which were created before refactoring the build scripts.
Therefore the dated content list makes a good regression test.