Serverspec: Fixes vars files for prod stage
The formatting in the vars/prod.yml file was off (and did not match the
format used by vars/staging.yml), and so was throwing an error during
the `rake` command. Updated the formatting and now the spectests start
correctly when running against prod hosts.

Also fixes the prod tests for Apache to use the newest headers in the
regex. Added an explantory comment with PR details in case the old
headers need to be checked for some reason. Prod instances aren't
actually using the new headers yet, so it's worth preserving the
information obviously for now.