Validates control field vars in build-debian-package role
Certain control fields are required for Debian packages, according to
the docs. Let's loop through those required fields and fail if any are
missing. For the "recommended" fields, we can print a debug message
about their recommended status, then proceed merrily along with the
build. Docs here:

  https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-controlsyntax

We're using the parameters for binary packages, not source packages,
mostly out of deference to how the packages have been built in the past.
At a later date we can evaluate adapting the template used.