Serverspec: Adds regression tests for securedrop-app-code fields
The build-debian-package role uses Conflicts and Replaces in the
template for DEBIAN/control, but the fields are written to the file even
if the values are blank. This doesn't seem to be a problem in terms of
control file syntax, but it's inelegant and could potentially cause
problems down the road. Safer to only write the fields that have
associated values.