Recommends line folding syntax for multiline YAML description
The `dpkg-deb` command doesn't like empty lines in the control file, and
will throw an error about too many fields if it encounters one. The most
likely way for an extra newline to creep in is via multiline YAML, using
the '>' operator, which appends a trailing newline to the concatenated
string. Instead, recommend the use of the line folding operator '>-',
which does not append a trailing newline.

Also updates wrapper roles to use the '>-' operator. It's possible that
the build-debian-package role could include tasks to clean up the
description field and strip trailing newlines, but it seems saner to
recommend sound practice than to munge data silently.