Serverspec: Disables regression testing for securedrop-app-code
The build-debian-package role is rather lax about permissions; it simply
removes write access to group and other. Certainly it's possible to
expose that functionality via a variable, but that doesn't seem
necessary right now. In the context of the securedrop-app-code package,
both the Ansible playbooks and the postinst script are fairly strict
about the permissions and ownership. Think more on how best to test for
regressions when building the app code Debian package.

Since this commit only affects the Serverspec test suite, I'm leaving it
in the history for posterity, rather than squashing it via rebase.