Disables /vagrant synchronized folder for build VM
The build-debian-package role now uses `fetch` to retrieve the built
Debian files, and `synchronize` to push up directories of files
recursively. Therefore there's no longer need to mount the /vagrant
shared directory, and in fact leaving it enabled slows down the CI tests
unnecessarily. It's now disabled, same as for the staging and prod hosts.