Adds "provider" variable in Vagrantfile
It's necessary to override a few vars depending on the backend provider,
such as VirtualBox or DigitalOcean. Mostly this is due to network
interfaces for OSSEC registration, as well as disabling
grsecurity-patched kernels on DigitalOcean.

In addition to checking for the `--provider` argument and the
environment variable `VAGRANT_DEFAULT_PROVIDER`, we'll need to walk the
machine directores where Vagrant stores this information.