Serverspec: Updates network interface lookups
The spectests were previously parsing the output of `hostname -I`, which
doesn't guarantee a reliable order of the IP addresses returned.
Therefore we cannot assume the first IP address belongs to eth0. Let's
update the tests to be smarter about interface handling, just as the
group_vars now smartly determine the iface based on provider.