Appends SSH args on non-Tails hosts
Tails will automatically route the SSH connections over Tor, same as it
does with all other traffic by default. In testing environments, however,
e.g. developer workstations, the most likely scenario is that Tor will
be running as a local service and SSH should be proxied through it.

Performing the Tails detection via output of `lsb_release`, which is the
same strategy Ansible uses when gathering hostfacts, but hostfacts are
not available during inventory construction.