Removes static inventory updates from install.sh script
The `install.sh` script bootstraps the Tails environment for the Admin
and Journalist Workstations, and on the Admin Workstation, it previously
edited the static inventory file to place the Onion URLs in the host
address field. A more durable solution is the dynamic inventory script
that reads these values from the local directory when trying to connect
to the servers, so now we can remove the difficult-to-test bash config
lines editing the old static inventory script.

Also updates the docs that referenced the old static inventory file
updating logic, since it's been removed.