Raises main stack gap 64KB -> 1MB via sysctl
We're already setting this value via the `securedrop-grsec` metapackage,
as described in #1861 and implemented in the grsec repo [0]. Let's also
ensure it at install time by setting it directly along with the other
sysctl options.

Included a config test so that test suite will fail if the sysctl
doesn't stick during CI or local development.

[0] https://github.com/freedomofpress/ansible-role-grsecurity/pull/100