Uses docker-compose for bootstrapping dev containers
The build logic isn't smartly dependent, so the first build will
needlessly redownload all the apt packages, which is the longest
part of the build process.

Creates shared data volume between app containers

Need to use the same database directory, since we don't have a service
container for the database (it's just sqlite).