Skip CircleCI if branch prefixed with `docs-`
Using the CircleCI branch filter settings [0] to skip running the
staging build environment if-and-only-if the PR branch name is prefixed
with `docs-`. Utilization of this feature is opt-in, so it will require
training regular contributors to make use of it, but should save 30-45
minutes for CI runs for those who do.

Updated the developer docs to make mention of the special branch prefix,
so taht first-time contributors have a shot at discovering it
themselves.

Closes #2132.

[0] https://circleci.com/docs/2.0/configuration-reference/#full-example