Removes incompatible versions of Firefox for app-test
Uninstalls any versions of Firefox other than the carefully tested
compatible version (to match pinned Selenium). The previous config
optimistically assumed that no Firefox was already provided on the
system, which is true of local VMs, but not true of Travis CI.

Moves Firefox tasks to separate task list

We were overloading the xvfb setup task list with firefox logic,
so let's break out the Firefox-specific tasks before changing them.
They're about to get a bit more complicated, in order to ensure that
we're using the correct version of Firefox, and no conflicting version
exists on the target system.