Removes deprecated Trusty image for Travis
Falling back to the now-stable new Trusty images, which showed
problems with Selenium. We must use the newer Trusty image, since the
deprecated version is scheduled for retirement on 2017-09-01 [0].

Changes will surely be required to our CI config for Travis in order
to use the new image, given the problems we saw on 2017-06-21 when the
default image switched to the new Trusty version.

[0] https://blog.travis-ci.com/2017-06-21-trusty-updates-2017-Q2-launch