Hardcodes Firefox binary path for Selenium in Travis
This shouldn't be necessary! Selenium by default uses `which` on Linux
platforms to figure out where the Firefox binary is located. For some
reason, that logic isn't working in Travis on the new Trusty image,
but hardcoded the path helps.

Made sure to hardcode the path only in Travis. Would prefer a single
codepath that works regardless of CI platform (and also in the local
development VM).