Stop running page-layout checks in Travis
The `--page-layout` flag for pytest activates the page-layout tests by
marking them via the pytest initialization process in conftest.py.
For reasons I have not discovered yet, the page-layout tests fail on the
new Travis Trusty image, but not on the old. The app unit tests, and
even the Selenium-based functional tests, are passing, however.

Also added a dump of the Selenium Firefox log at the end of the app test
run, so that failed CI runs are more informative. Note that the failing
page-layout tests hung indefinitely several times during testing, so
the app log dump may not always work. Still, better to have it sometimes
than never.