Ensures xvfb service is running in app-test role
The app-test role included a "handler" for starting xvfb, which breaks
idempotence in the edge case of an error being encountered in the
playbook run before the handler has fired: the service is never started,
and subsequent runs of the playbook don't modify files, so the handler
is never called.

The best practice approach via Ansible is to have a task explicitly
start the service in question, then fire a handler for "restarted"
changes whenever a config file is modified.