Uses sudo in testinfra tests for pip dependencies
The new Travis Trusty image includes a lot of Python packages that are
installed either in a virtualenv, provided by Travis, or globally, via
our Ansible scripts. Since the Ansible scripts use sudo, we must
similarly use sudo to check for their presence, otherwise we're just
listing the contents of the virtualenv.

Testinfra does not support use of sudo with the LocalCommand fixture,
but *does* allow it when target localhost implicitly via direct
invocation.

The pip version mismatches were discovered by comparing the results of:

  * pip freeze -l
  * sudo pip freeze -l

in the Travis CI run.