Updates Firefox config tests for Travis CI
The new Travis Trusty image provides a version of Firefox located in the
home directory, and added to the PATH for the Travis user. Therefore the
package installed=yes state check will fail. Instead, let's only check
for apt presence in the development VM, but in both contexts, confirm
that "firefox --version" shows the proper version, i.e. 46.0.1.