CLEANUP: Disables Travis Python customizations
Setting language=generic to avoid use of virtualenv. We do not use a
virtualenv for configuring SecureDrop in production, nor in the
Development VM, which the Travis environment (rather crudely) emulates.

Removing Python-specific Travis helper definitions, such as pinning
version 2.7 (which is already the default), and permitting system
packages in the virtualenv (which is no longer used).

Perhaps most cryptically, we're now modifying the PATH to remove
references to /opt/python. Travis includes a few bin directories in that
location, which wreak havoc with our app test suite.