Stop running tests as root in Travis
We've shaken out most of the pip dependencies woes, so we can finally
act like humans again and run the tests with normal user privileges.
Includes the `--page-layout` option and all tests are passing.

Includes a comment about the necessity to execute the test suite
from a very specific directory. The problem is temporary, and tracked
in a dedicated issue. We can live with it for now. Mostly I'm satisfied
that a thorough investigation of the Travis config shook out bugs
such as this one.