Stops hardcoding Firefox binary path for Selenium in Travis
Since moving to the new Ubuntu Trusty image in Travis, Firefox 46.0.1 is
now installed automatically, within `~travis`. Since the tests were
previously being executed as root, the PATH modification for the travis
user wasn't exposed to the Selenium webdriver discovery logic, since
even that was being executed as root.

Now that we're running the app test suite with normal user privileges,
we can expect Selenium to do the right thing in ferretting out where the
Firefox binary is located, based on `which`.

Revert "Hardcodes Firefox binary path for Selenium in Travis"

This reverts commit 3c4b17240ba930d758a9145dfb82c1902f7755ef.