Adds docker-compose for dev env
Sprinkles in a base dockerfile for use in development (should be reused
during tests when those are containerized).

Sets SECUREDROP_ENV=dev in containers
Necessary to ensure Flask serves static assets; in "prod" env, Flask
uses X-Send-File, to be handled by the webserver.

Not sure whether it's necessary to build the sass inside the Dockerfile,
maybe we could move that asset collection logic elsewhere. Fine for now,
though.