Makes Redis server host configurable
In the development and app-prod deployment story, the redis server is
always run on "localhost", so the value never needed to be changed. In
the container story, however, "localhost" doesn't mean the same thing,
so we'll need to provide customizations via env var to point to the
"redis" server name.