Supports Qubes for running the staging VMs
Using the "delegated" driver for Molecule, meaning that create
and destroy actions must be handled explicitly via the playbooks
within the scenario (rather than leaning on a convenient API).

There are a few shortcomings with the current implementation:

  * become pass is hardcoded (due to template VMs; should be changed)
  * VMs only halt, not start, when rebooted
  * the create/destroy logic is task-based, rather than module-based

we should consider writing a "qubes_vm" Ansible module, so the VMs
can be managed via e.g. state=started, state=absent, etc. The catch,
however, is that the `qubesadmin` Python library is only installable
via apt/yum, meaning that the libs won't be available inside the
virtualenv without a sys.path hack.